+++ jme3-desktop/src/main/java/jme3tools/navigation/MapModel3D.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009-2012 jMonkeyEngine
+ * Copyright (c) 2009-2020 jMonkeyEngine
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -120,7 +120,7 @@ public class MapModel3D {
      */
     public void calculateMinutesPerWorldUnit(double mapWidthInLongitude) {
         // Multiply mapWidthInLongitude by 60 to convert it to minutes.
-        minutesPerWorldUnit = (mapWidthInLongitude * 60) / (double) worldWidth;
+        minutesPerWorldUnit = (mapWidthInLongitude * 60) / worldWidth;
     }
 
     /**
